refactor: transfered unit test of queryFilter from location to common [10022]
